LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-scope-expected.txt


FAIL Descendant can attach to deferred timeline assert_equals: expected "100px" but got "0px"
PASS Deferred timeline with no attachments
PASS Inner timeline does not interfere with outer timeline
PASS Deferred timeline with two attachments
FAIL Dynamically re-attaching assert_equals: expected "100px" but got "0px"
FAIL Dynamically detaching assert_equals: expected "0px" but got "100px"
FAIL Removing/inserting element with attaching timeline assert_equals: expected "100px" but got "0px"
FAIL Ancestor attached element becoming display:none/block assert_equals: expected "100px" but got "0px"
FAIL A deferred timeline appearing dynamically in the ancestor chain assert_equals: expected "0px" but got "100px"
PASS Animations prefer non-deferred timelines
